# backend/docs/swagger.yaml
basePath: /
definitions:
  model.ErrorResponse:
    properties:
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  model.PathData:
    properties:
      grpc:
        type: string
      upgrade:
        type: string
      ws:
        type: string
    type: object
  model.PayloadWSData:
    properties:
      payloadcdn:
        type: string
      payloadwithpath:
        type: string
    type: object
  model.PortData:
    properties:
      grpc:
        items:
          type: string
        type: array
      none:
        items:
          type: string
        type: array
      tls:
        items:
          type: string
        type: array
      udpgw:
        items:
          type: string
        type: array
    type: object
  model.SSHData:
    properties:
      city:
        type: string
      expired:
        type: string
      hostname:
        type: string
      isp:
        type: string
      password:
        type: string
      payload_ws:
        $ref: '#/definitions/model.PayloadWSData'
      port:
        $ref: '#/definitions/model.PortData'
      username:
        type: string
    type: object
  model.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  model.TrojanData:
    properties:
      city:
        type: string
      expired:
        type: string
      hostname:
        type: string
      isp:
        type: string
      link:
        $ref: '#/definitions/model.XrayLinkData'
      password:
        type: string
      path:
        $ref: '#/definitions/model.PathData'
      port:
        $ref: '#/definitions/model.PortData'
      username:
        type: string
    type: object
...
  model.XrayLinkData:
    properties:
      grpc:
        type: string
      none:
        type: string
      tls:
        type: string
      upnone:
        type: string
      uptls:
        type: string
    type: object
info:
  contact: {}
  description: Selamat datang di dokumentasi API Neko Tunneling. Halaman ini berisi
    referensi endpoint lengkap untuk mempermudah integrasi VPS dengan Web Panel, skrip
    automasi, maupun Bot Telegram.
  title: Neko Tunneling API
  version: "1.0"
paths:
  /vps/create/ssh:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Username akun
        in: formData
        name: username
        required: true
        type: string
      - description: Pasword akun (akan dibuat random jika kosong)
        in: formData
        name: password
        type: string
      - description: Limit IP (unlimited jika kosong)
        in: formData
        name: limit_ip
        type: integer
      - description: Limit Quota (unlimited jika kosong)
        in: formData
        name: limit_quota
        type: integer
      - description: Durasi masa aktif (hari)
        in: formData
        name: expired
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.SSHData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Akun SSH
      tags:
      - Create Akun
  /vps/create/trojan:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Username akun
        in: formData
        name: username
        required: true
        type: string
      - description: Pasword akun (akan dibuat random jika kosong)
        in: formData
        name: password
        type: string
      - description: Limit IP (unlimited jika kosong)
        in: formData
        name: limit_ip
        type: integer
      - description: Limit Quota (unlimited jika kosong)
        in: formData
        name: limit_quota
        type: integer
      - description: Durasi masa aktif (hari)
        in: formData
        name: expired
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.TrojanData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Akun Trojan
      tags:
      - Create Akun
...
  /vps/trial/ssh:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Durasi Trial (menit)
        in: formData
        name: expired
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.SSHData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Trial Akun SSH
      tags:
      - Trial Akun
  /vps/trial/trojan:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Durasi Trial (menit)
        in: formData
        name: expired
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.TrojanData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Trial Akun Trojan
      tags:
      - Trial Akun
...
securityDefinitions:
  BearerAuth:
    description: Masukkan token Anda saja (Awalan "Bearer " akan ditambahkan otomatis)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"